@@ -11,16 +11,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "17.3.12",
+    "@angular/animations": "18.2.13",
     "@angular/cdk": "^17.3.10",
-    "@angular/common": "17.3.12",
-    "@angular/compiler": "17.3.12",
-    "@angular/core": "17.3.12",
-    "@angular/forms": "17.3.12",
+    "@angular/common": "18.2.13",
+    "@angular/compiler": "18.2.13",
+    "@angular/core": "18.2.13",
+    "@angular/forms": "18.2.13",
     "@angular/material": "^17.3.10",
-    "@angular/platform-browser": "17.3.12",
-    "@angular/platform-browser-dynamic": "17.3.12",
-    "@angular/router": "17.3.12",
+    "@angular/platform-browser": "18.2.13",
+    "@angular/platform-browser-dynamic": "18.2.13",
+    "@angular/router": "18.2.13",
     "ng-flex-layout": "^17.3.7-beta.1",
     "ngx-custom-material-file-input": "^4.0.0",
     "rxjs": "7.5.6",
@@ -28,14 +28,14 @@
     "zone.js": "^0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "17.3.17",
+    "@angular-devkit/build-angular": "18.2.20",
     "@angular-eslint/builder": "14.0.2",
     "@angular-eslint/eslint-plugin": "14.0.2",
     "@angular-eslint/eslint-plugin-template": "14.0.2",
     "@angular-eslint/schematics": "^18.4.3",
     "@angular-eslint/template-parser": "14.0.2",
-    "@angular/cli": "17.3.17",
-    "@angular/compiler-cli": "17.3.12",
+    "@angular/cli": "18.2.20",
+    "@angular/compiler-cli": "18.2.13",
     "@types/jasmine": "4.0.3",
     "@typescript-eslint/eslint-plugin": "5.33.1",
     "@typescript-eslint/parser": "5.33.1",
@@ -46,6 +46,6 @@
     "karma-coverage": "2.2.0",
     "karma-jasmine": "5.1.0",
     "karma-jasmine-html-reporter": "2.0.0",
-    "typescript": "5.2"
+    "typescript": "5.5.4"
   }
 }
